woocommerce/mcp/requirements.txt
google-generativeai
python-dotenv
requests
mcp